Move all remaining model mapping in charge-processor to services directory
