+++ src/modules/billing/services/invoice-licences-service.js
@@ -1,0 +1,35 @@
+const Licence = require('../../../lib/models/licence');
+const InvoiceLicence = require('../../../lib/models/invoice-licence');
+
+const { mapCRMAddressToModel } = require('./address-service');
+const { mapCRMCompanyToModel } = require('./companies-service');
+const { mapCRMContactToModel } = require('./contacts-service');
+
+/**
+ * Maps a charge version from the charge processor to a Licence instance
+ * @param {Object} chargeVersion
+ * @return {Licence}
+ */
+const mapLicence = chargeVersion => {
+  const licence = new Licence();
+  licence.licenceNumber = chargeVersion.licenceRef;
+  return licence;
+};
+
+/**
+ * Maps a row of data from the charge processor to an InvoiceLicence instance
+ * @param {Object} data - processed charge version
+ * @return {InvoiceLicence}
+ */
+const mapChargeRowToModel = data => {
+  const invoiceLicence = new InvoiceLicence();
+  invoiceLicence.licence = mapLicence(data.chargeVersion);
+  invoiceLicence.company = mapCRMCompanyToModel(data.licenceHolder.company);
+  invoiceLicence.address = mapCRMAddressToModel(data.licenceHolder.address);
+  if (data.licenceHolder.contact) {
+    invoiceLicence.contact = mapCRMContactToModel(data.licenceHolder.contact);
+  }
+  return invoiceLicence;
+};
+
+exports.mapChargeRowToModel = mapChargeRowToModel;
