+++ src/modules/billing/services/invoice-service.js
@@ -1,11 +1,16 @@
 'use strict';
 
-const { first } = require('lodash');
+const { first, uniqBy } = require('lodash');
 
 const repos = require('../../../lib/connectors/repository');
+
+// Services
 const transactionsService = require('./transactions-service');
 const invoiceAccountsService = require('./invoice-accounts-service');
+const addressService = require('./address-service');
+const invoiceLicencesService = require('./invoice-licences-service');
 
+// Models
 const Invoice = require('../../../lib/models/invoice');
 const InvoiceAccount = require('../../../lib/models/invoice-account');
 const InvoiceLicence = require('../../../lib/models/invoice-licence');
@@ -143,5 +148,56 @@ const getInvoiceForBatch = async (batchId, invoiceId) => {
   }
 };
 
+const getInvoiceAccountNumber = row => row.invoiceAccount.invoiceAccount.invoiceAccountNumber;
+
+/**
+ * Maps output data from charge processor into an array of unique invoice licences
+ * matching the invoice account number of the supplied Invoice instance
+ * @param {Invoice} invoice - invoice instance
+ * @param {Array} data - processed charge versions
+ * @return {Array<InvoiceLicence>}
+ */
+const mapInvoiceLicences = (invoice, data) => {
+  // Find rows with invoice account number that match the supplied invoice
+  const { accountNumber } = invoice.invoiceAccount;
+  const filtered = data.filter(row => getInvoiceAccountNumber(row) === accountNumber);
+  // Create array of InvoiceLicences
+  const invoiceLicences = filtered.map(invoiceLicencesService.mapChargeRowToModel);
+  // @todo attach transactions to InvoiceLicences
+  // Return a unique list
+  return uniqBy(invoiceLicences, invoiceLicence => invoiceLicence.uniqueId);
+};
+
+/**
+ * Given an array of data output from the charge processor,
+ * maps it to an array of Invoice instances
+ * @param {Array} data
+ * @return {Array<Invoice>}
+ */
+const mapChargeDataToModels = data => {
+  // Create unique list of invoice accounts within data
+  const rows = uniqBy(
+    data.map(row => row.invoiceAccount),
+    row => row.invoiceAccount.invoiceAccountId
+  );
+
+  // Map to invoice models
+  return rows.map(row => {
+    const invoice = new Invoice();
+
+    // Create invoice account model
+    invoice.invoiceAccount = invoiceAccountsService.mapCRMInvoiceAccountToModel(row.invoiceAccount);
+
+    // Create invoice address model
+    invoice.address = addressService.mapCRMAddressToModel(row.address);
+
+    // Create invoiceLicences array
+    invoice.invoiceLicences = mapInvoiceLicences(invoice, data);
+
+    return invoice;
+  });
+};
+
 exports.getInvoicesForBatch = getInvoicesForBatch;
 exports.getInvoiceForBatch = getInvoiceForBatch;
+exports.mapChargeDataToModels = mapChargeDataToModels;
