+++ test/modules/billing/services/batch-service.js
@@ -5,9 +5,9 @@ const {
 } = exports.lab = require('@hapi/lab').script();
 const { expect } = require('@hapi/code');
 
-const { Batch, Invoice, InvoiceAccount, Address, Company } = require('../../../../../src/lib/models');
-const Contact = require('../../../../../src/lib/models/contact-v2');
-const modelMapper = require('../../../../../src/modules/billing/service/charge-processor/model-mapper');
+const { Batch, Invoice, InvoiceAccount, Address, Company } = require('../../../../src/lib/models');
+const Contact = require('../../../../src/lib/models/contact-v2');
+const batchService = require('../../../../src/modules/billing/services/batch-service');
 
 const BATCH_ID = '6556baab-4e69-4bba-89d8-7c6403f8ac8d';
 
@@ -72,13 +72,13 @@ const createData = () => [{
   }
 }];
 
-experiment('modules/billing/service/charge-processor/model-mapper.js', () => {
-  experiment('modelMapper', () => {
+experiment('modules/billing/services/batch-service', () => {
+  experiment('.mapChargeDataToModel', () => {
     let data, result, invoice;
 
     beforeEach(async () => {
       data = createData();
-      result = modelMapper.modelMapper(BATCH_ID, data);
+      result = batchService.mapChargeDataToModel(BATCH_ID, data);
     });
 
     test('should return a batch with the correct ID', async () => {
